Comiezno carga asincronica.
Subo estos cambios por las dudas.
