--- Assets/00_Scenes/2_SecondLevel.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9112cdc8ba0d59149bbe776f7977b14c
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
