+++ Assets/00_Scenes/2_SecondLevel/Bongo Size.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 053db1a643af88c49ac3e6bcfb936835
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
