+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_EndZone.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e5935302cb9df3941aba6de290990507
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
