+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone2.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 156054344ca8a0c44b8dfae0786f8cfe
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
