+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone3.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 42634b663ead12248a85254fdedb4357
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
