+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone4.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e23c1ce3229bda14f9bac97641314880
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
