+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone5.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 543e01dabdb74cf4ba68c579f957d922
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
