+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone6.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cef9622f81fb82e42a340585762641ef
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
