+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone7.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d06141979186e89459a9c51017c664c2
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
