+++ Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_ZoneCollectable.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff1fc94c9db348b4ab5705788ece2016
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
