--- Assets/00_Scenes/3_ThirdLevel.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 99b8e980b7a9d7f48b3a2c4b47e84ada
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
