--- Assets/03_Scripts/00_Managers/LoadSceneCollider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6ad2d94ee3b8aa046998c10b2cabe2d0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
