+++ Assets/03_Scripts/00_Managers/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be871650991dc604fb574e903ced86dc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
