+++ Assets/03_Scripts/00_Managers/Scenes/AsyncLoad.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64fde2a9adf31a44ba75f20c5bad7560
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
