+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,9 +5,6 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
-    path: Assets/00_Scenes/NIVELES VIEJOS/old_0_MainMenuBook.unity
-    guid: c251d36eb9782814bb5be112d482cd5a
   - enabled: 1
     path: Assets/00_Scenes/0_MainMenuBook.unity
     guid: 3f64f023612628544a23471dd7501f22
@@ -18,7 +15,7 @@ EditorBuildSettings:
     path: Assets/00_Scenes/1_FirstLevel.unity
     guid: 7a47390ffdf19bd42854298f37131438
   - enabled: 1
-    path: Assets/00_Scenes/2_SecondLevel.unity
+    path: Assets/00_Scenes/2_SecondLevel/2_SecondLevel.unity
     guid: 9112cdc8ba0d59149bbe776f7977b14c
   - enabled: 1
     path: Assets/00_Scenes/3_ThirdLevel.unity
@@ -29,13 +26,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/00_Scenes/1_ZoneColl.unity
     guid: 60e1125f4d2f1ee40a69d7d30a9b561f
-  - enabled: 0
-    path: Assets/00_Scenes/4_FourthLevel.unity
-    guid: 5ef4df59971e6584ab42304e1fce1618
-  - enabled: 0
-    path: Assets/00_Scenes/NIVELES VIEJOS/3_OldThirdLevel.unity
-    guid: 5ed5fa117d345074f9dc4f8ed5c9daef
-  - enabled: 0
-    path: Assets/00_Scenes/5_FifthLevel.unity
-    guid: 2a8a6a432fd3d1243846a49df638fbad
+  - enabled: 1
+    path: Assets/00_Scenes/2_SecondLevel/Bongo Size/2B_Zone2.unity
+    guid: 156054344ca8a0c44b8dfae0786f8cfe
   m_configObjects: {}
